+++ src/utils.js
@@ -32,7 +32,7 @@ let getProvider = _.curry(
     )
 )
 
-let getChildren = F.cascade(['children', 'data.children'])
+let getChildren = F.cascade(['children', 'items', 'data.items'])
 let getRelevantFilters = _.curry((groupCombinator, Path, group) => {
   if (!_.includes(group.key, Path))
     // If we're not in the path, it doesn't matter what the rest of it is