+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ﻿### 0.1.0
-- Using `children` instead of `items`.
+- Supporting `children`.
 
 ### 0.0.5
 - Ecosystem And Resources readme update (version bump for npm pubishing)